substrate_runtime

Type Alias AllPalletsWithSystem

pub type AllPalletsWithSystem = (System, Timestamp, Balances, Sudo, TransactionPayment, Revive);

All pallets included in the runtime as a nested tuple of types.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 0 bytes