substrate_runtime

Type Alias ReviveConfig

pub type ReviveConfig = GenesisConfig<Runtime>;

Aliased Type

pub struct ReviveConfig {
    pub mapped_accounts: Vec<AccountId32>,
    pub accounts: Vec<Account<Runtime>>,
    pub debug_settings: Option<DebugSettings>,
}

Fields
mapped_accounts: Vec<AccountId32>

List of native Substrate accounts (typically AccountId32) to be mapped at genesis block, enabling them to interact with smart contracts.

accounts: Vec<Account<Runtime>>

Account entries (both EOAs and contracts)

debug_settings: Option<DebugSettings>

Optional debugging settings applied at genesis.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 56 bytes