Crate substrate_runtime

Modules

api

currency

genesis_config_presets

Provides getters for genesis configuration presets.

Structs

ChainId

CodeHashLockupDepositPercent

EthExtraImpl

Default extensions applied to Ethereum transactions.

ExistentialDeposit

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

TransactionByteFee

Version

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

AVERAGE_ON_INITIALIZE_RATIO 🔒

We assume that ~10% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

MAXIMUM_BLOCK_WEIGHT 🔒

We allow for 2 seconds of compute with a 6 second average block time, with maximum proof size.

NORMAL_DISPATCH_RATIO 🔒

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

RUNTIME_API_VERSIONS

VERSION

The runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Traits

InherentDataExt 🔒

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balance of an account.

Balances

Provides the ability to keep track of balances.

BalancesConfig

Block

Block type as expected by this runtime.

Executive 🔒

Hash

Hash type used by the runtime.

Revive

Provides the ability to execute Smart Contracts.

ReviveConfig

Sudo

Provides a way to execute privileged functions.

SudoConfig

System

Mandatory system pallet that should always be included in a FRAME runtime.

SystemConfig

Timestamp

Provides a way for consensus systems to set and check the onchain time.

TransactionPayment

Provides the ability to charge for extrinsic execution.

TransactionPaymentConfig

TxExtension 🔒

The transaction extensions that are added to the runtime.

UncheckedExtrinsic